# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "summary": "Sign in",
        "description": "Checks email and password. Returns a short-lived access token (send it as `Authorization: Bearer`) and sets the httpOnly `wms_refresh` cookie for POST /auth/refresh.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "`{ accessToken, user: SessionUser }`"
          },
          "401": {
            "description": "`invalid_credentials`"
          },
          "429": {
            "description": "`rate_limited`"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "summary": "New access token from the refresh cookie",
        "description": "Reads `wms_refresh`, extends the session and returns a fresh access token.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "`{ accessToken, user: SessionUser }`"
          },
          "401": {
            "description": "`unauthenticated`: the session ended or expired"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/claims": {
      "get": {
        "operationId": "Claims_list",
        "summary": "Claims (A09, DL07)",
        "description": "`q` matches id, serial, RMA number and brand. Dealers and distributors get their own, read-only.",
        "parameters": [
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "PAID",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated<ClaimView>"
          }
        },
        "tags": [
          "claims"
        ]
      }
    },
...
    "/api/claims/{id}": {
      "get": {
        "operationId": "Claims_get",
        "summary": "Claim with the job result as evidence (A10)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ClaimView"
          }
        },
        "tags": [
          "claims"
        ]
      }
    },
    "/api/claims/{id}/transitions": {
      "post": {
        "operationId": "Claims_transition",
        "summary": "Move a claim to its next status",
        "description": "`submit` (needs amount > 0; rmaNumber optional) writes the OEM submission; `mark_paid` posts to Finance. `422` validation.amount / validation.reasonRequired; `409 invalid_transition`.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "submit",
                      "approve",
                      "reject",
                      "mark_paid"
                    ]
                  },
                  "rmaNumber": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ClaimView"
          }
        },
        "tags": [
          "claims"
        ]
      }
    },
    "/api/complaints": {
      "get": {
        "operationId": "Complaints_list",
        "summary": "Complaints in the caller's scope (A07, DL07, CU05)",
        "description": "`q` matches id, serial, customer and dealer. Newest first by default.",
        "parameters": [
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "CUSTOMER",
                "DEALER",
                "ADMIN"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NEW",
                "WITH_SERVICE",
                "RESOLVED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated<ComplaintView>"
          }
        },
        "tags": [
          "complaints"
        ]
      },
      "post": {
        "operationId": "Complaints_create",
        "summary": "Raise a complaint (CU04, DL06)",
        "description": "The source follows the caller's role; the server decides the entitlement. `422` validation.describeFault; `409 not_registered`; `404` for a unit outside the caller's scope.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "unitSerial",
                  "description"
                ],
                "properties": {
                  "unitSerial": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 5
                  },
                  "attachmentIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ComplaintView"
          }
        },
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{id}": {
      "get": {
        "operationId": "Complaints_get",
        "summary": "Complaint with entitlement, job result and claim status (A08, CU05)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ComplaintView"
          }
        },
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{id}/send-to-service": {
      "post": {
        "operationId": "Complaints_sendToService",
        "summary": "Send to the service system (A08)",
        "description": "NEW -> WITH_SERVICE; writes the outbound service request. `409 already_sent`.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ComplaintView"
          }
        },
        "tags": [
          "complaints"
        ]
      }
    },
...
  },
  "info": {
    "title": "Warranty Management API",
    "description": "Units with part-wise warranties, registrations from every channel, complaints, service hand-off, manufacturer claims and the integration log. Contract: frontend/docs/api-contract.md.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wms_refresh"
      }
    },
...
  }
}